Implement example page with Next getServerSideProps interception
The `/pizza` page takes a long time to load because of the long
`getServerSideProps`. This hinders user experience when navigating to
that page in the browser. The user has to way for that page to finish
rendering before it is shown. It seems like the app is stuck, there is
no other loading indicator.

When the Next `getServerSideProps` interception is enabled, the browser
shows the route immediately during client-side navigation. It is up to
the page to then render the loading state and fetch the appropriate data
in the browser.

This is better than opting out of `getServerSideProps` because the
`/pizza` page can be fully rendered in SSR with the data already
present. This improves SEO.